backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Async HTTP and File Operations
httpx>=0.26.0
aiofiles>=23.2.1

# Environment Variables
python-dotenv>=1.0.0

# Caching and Task Queue
redis>=5.0.1
celery>=5.3.4

# Storage (S3/MinIO)
boto3>=1.34.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code Quality
ruff>=0.1.14
mypy>=1.8.0

# Demo Data Generation
faker>=22.0.0
